Major Update
This update is almost a complete overwrite of all the templates in the
entire system.   It contains a large number of changes to the language
auto detection and it's use of cookies along with many small
improvements to the main .htaccess file and the /ccmstpl/index.tpl file
which now has all of it's content relocated to the /ccmstpl/examples/
folder.