v0.4 updates, header changes, template parser changes, template rename
- Removed references to .htm, .tpl, .txt, .xml and .xsl from parser.
- Disabled call to CCMS_setContentTypeHeader() function.
- Relocated header declaration (Content-Type: text/html) to only
function when the ccms_tpl is associated with .html templates.  ccms_tpl
calls to .php templates are now expected to generate their own headers
from now on.  This puts the responsibility back into the hands of
website developers to determine the function of the .php template.
- Addition of underscores to define('TPL', '/^[a-z-_\pN\.\/]*\z/i') for
templates that contain _ in their names.
- Removal of several 'Debug' code snippets.
- Updates to many ifelse statements in the /ccmspre/index.php template
to remove decremented .htm, .tpl, .txt, .xml and .xsl references.
- Updates to .htaccess files containing decremented .htm, .tpl, .txt,
.xml and .xsl references.
- Many smaller example template name changes and minor content updates.