Added server side caching and other minor updates.
- Added new server side caching.  Custodian CMS already automatically
enables caching of .html pages on the client side, now you can enable
and set the length of time (in minutes) you want it to store a copy of a
generated .html page in the database.  .php templates are not cached in
order to guarantee developer control of custom headers.  Server cache
setting are not turned on by default to help with initial site
development but can be enable and configure in the /ccmspre/config.php.
Eg:
$CFG["CACHE"] = 0; // off (Default)
$CFG["CACHE"] = 1; // on
$CFG["CACHE_EXPIRE"] = 360; // 6 hours  (Default)
$CFG["CACHE_EXPIRE"] = 720; // 12 hours
- The /setup.php file was updated with all the new server-side cache
settings.
- “ccms_cache” was added to the database to hold all cached templates.
- All instances of // commenting were removed from the example templates
and replaced with /* ...text comment…*/ to help with the new
CCMS_html_min() function.
- Removed browser.php from /ccmstpl/examples/.